输入法测试

+++ app/src/main/java/com/lenovo/sunzh/permissionsdm/utils/PermissionUtil.java
@@ -31,7 +31,7 @@ public class PermissionUtil {
             try {
                 Class clazz = AppOpsManager.class;
                 Method method = clazz.getDeclaredMethod("checkOp", int.class, int.class, String.class);
-                return AppOpsManager.MODE_ALLOWED == method.invoke(manager, op, Binder.getCallingUid(), context.getPackageName());
+                return AppOpsManager.MODE_ALLOWED == (int) method.invoke(manager, op, Binder.getCallingUid(), context.getPackageName());
             } catch (Exception e) {
                 e.printStackTrace();
                 Log.e("TAG", Log.getStackTraceString(e));
